+++ tests/database/factories/TagFactory.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Happytodev\Blogr\Tests\Database\Factories;
+
+use Happytodev\Blogr\Models\Tag;
+use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Support\Str;
+
+/**
+ * @extends \Illuminate\Database\Eloquent\Factories\Factory<\Happytodev\Blogr\Models\Tag>
+ */
+class TagFactory extends Factory
+{
+    protected $model = Tag::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array<string, mixed>
+     */
+    public function definition(): array
+    {
+        $name = fake()->word();
+
+        return [
+            'name' => ucfirst($name),
+            'slug' => Str::slug($name),
+        ];
+    }
+}