+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 All notable changes to `blogr` will be documented in this file.
 
-## [Unreleased]
+
+## [v0.3.0](https://github.com/happytodev/blogr/compare/v0.2.1...v0.3.0) - 2025-09-01
 
 ### 🚀 Feature